Update settings load order for celery settings, clean up docs
This uses the supervisord config as an example illustration, instead of
burying the commands in the docs.
