+++ contrib/supervisord.conf
@@ -17,14 +17,6 @@ password = 🍟
 [rpcinterface:supervisor]
 supervisor.rpcinterface_factory = supervisor.rpcinterface:make_main_rpcinterface
 
-[program:redis]
-command = redis-server
-directory = ../
-priority = 0
-startretries = 0
-stderr_logfile = /dev/stderr
-stderr_logfile_maxbytes = 0
-
 [program:main]
 command = python3 manage.py runserver 8000
 directory = ../
@@ -33,10 +25,23 @@ stopasgroup = True
 stderr_logfile = /dev/stderr
 stderr_logfile_maxbytes = 0
 
-[program:celery]
-command = python3 -m celery worker -A readthedocs.worker -Q default,celery,web,builder -l DEBUG -c 2
+[program:web]
+command = python3 -m celery worker -A readthedocs.worker -Q web -l DEBUG -c 1
+environment = DJANGO_SETTINGS_MODULE="readthedocs.settings.dev.celery.web"
 directory = ../
 priority = 100
 stopasgroup = True
 stderr_logfile = /dev/stderr
 stderr_logfile_maxbytes = 0
+
+[program:build]
+command = python3 -m celery worker -A readthedocs.worker -Q default,celery,builder -l DEBUG -c 1
+environment = DJANGO_SETTINGS_MODULE="readthedocs.settings.dev.celery.build"
+directory = ../
+priority = 100
+stopasgroup = True
+stderr_logfile = /dev/stderr
+stderr_logfile_maxbytes = 0
+
+[group:celery]
+programs = web,build
