+++ readthedocs/settings/dev/celery/web.py
